Consolidate .netcoreapp.cs files because System.Diagnostics.* projects is no longer cross-compiled
