+++ src/libraries/System.Diagnostics.TraceSource/tests/DefaultTraceListenerClassTests.cs
@@ -4,11 +4,14 @@
 
 using System.IO;
 using System.Reflection;
+using System.Reflection.Emit;
+using System.Linq;
+using Microsoft.DotNet.RemoteExecutor;
 using Xunit;
 
 namespace System.Diagnostics.TraceSourceTests
 {
-    public partial class DefaultTraceListenerClassTests : FileCleanupTestBase
+    public class DefaultTraceListenerClassTests : FileCleanupTestBase
     {
         private class TestDefaultTraceListener : DefaultTraceListener
         {
@@ -138,6 +141,44 @@ namespace System.Diagnostics.TraceSourceTests
             Assert.Equal(string.Empty, listener.LogFileName);
             listener.LogFileName = expectedLogFileName;
             Assert.Equal(expectedLogFileName, listener.LogFileName);
+        }
+
+        [Fact]
+        public void EntryAssemblyName_Default_IncludedInTrace()
+        {
+            RemoteExecutor.Invoke(() =>
+            {
+                var listener = new TestDefaultTraceListener();
+                Trace.Listeners.Add(listener);
+                Trace.TraceError("hello world");
+                Assert.Equal(Assembly.GetEntryAssembly()?.GetName().Name + " Error: 0 : hello world", listener.Output.Trim());
+            }).Dispose();
+        }
+
+        [Fact]
+        public void EntryAssemblyName_Null_NotIncludedInTrace()
+        {
+            RemoteExecutor.Invoke(() =>
+            {
+                MakeAssemblyGetEntryAssemblyReturnNull();
+
+                var listener = new TestDefaultTraceListener();
+                Trace.Listeners.Add(listener);
+                Trace.TraceError("hello world");
+                Assert.Equal("Error: 0 : hello world", listener.Output.Trim());
+            }).Dispose();
+        }
+
+        /// <summary>
+        /// Makes Assembly.GetEntryAssembly() return null using private reflection.
+        /// </summary>
+        private static void MakeAssemblyGetEntryAssemblyReturnNull()
+        {
+            typeof(Assembly)
+                .GetField("s_forceNullEntryPoint", BindingFlags.NonPublic | BindingFlags.Static)
+                .SetValue(null, true);
+
+            Assert.Null(Assembly.GetEntryAssembly());
         }
     }
 }
